Adding mock authentication implementation
I have added in a mock authentication function in
the App component this simply toggles the state of
authenticated=true if the username and password
are truth-y. This is passed as a prop to the
LoginForm component, which stores the username
and password as component states. Finally, the
login button on the LoginForm component has been
set to use the authentication function onClick.