Creating first testing file
I am using testing-library/react with Jest to create
a testing suite for the front-end. Due to where Jest
looks for the test files, I moved them into /src/.
Currently, App.test.js tests the App component,
seeing if it initially renders the typical parts of a
login screen and makes sure that it initially is has
blank username and password inputs.