Adding mock URL Shortener form
After the mock authentication process, I now have the App component render a new
component (src/components/ShortenerForm), which renders with the title of the app,
two text input boxes for a shortened URL and a full URL, and a button to add the URL to
the database. Currently, there is no functionality to this component. The testing suite
in src/App.test.js has been updated such that it tests to see if the parts of the new
page are rendered.

 Changes to be committed:
	modified:   src/App.test.js
	modified:   src/components/App/App.js
	new file:   src/components/ShortenerForm/ShortenerForm.css
	new file:   src/components/ShortenerForm/ShortenerForm.js