Adding Key-Value persistence test to test suite
Added in a fourth test into src/App.test.js that tests
to see if the Key-Value pair (short/full URLs) are
saved to the user's main page after they've been
authenticated and add the pair using the Shortener
Form. Currently, it doesn't pass (nor should it), but
that's because each test re-renders the App
component, which resets the mock authentication
to false. Thus, I feel like there needs to be a better
way to do these tests which requires thinking.