Created test suite for LoginForm
I created the first test for the LoginForm component
(in src/LoginForm.test.js) that simply tests if it initially renders with a blank username and password.
I eliminated an identical test in src/App.test.js that
did the same thing. This makes the different test
files more focused in their purpose.