Adding initial test for user adding URL
I added new test IDs to the inputs in UrlAdder so that I can use UrlForm.test.js
to add in a test where the user would add
a short/full URL pair. One test just simply
checks to see if a UrlListElement is now
in the UrlForm (which its not) and that
it contains the pair (which it does not).
This now triggers the development of the
UrlListElement component and some of
the base functionality.