Adding UrlAdder and UrlList functionality
Firstly, I discovered that UrlForm.test.js
needed to still be changed in order to
properly test everything (i.e. I couldn't
just render the UrlForm at the beginning
of the describe block) as this resulted in
improper behavior. I added in the list of
URL pairs to the UrlForm as a state, which
then gets passed to UrlList, which renders
the list as a series of UrlListElement
components. Finally, I added the ability
to add to the list using the addUrlPair
method in UrlForm, which passes it to the
UrlAdder, which then uses it as an onClick
event function for the add button.
Currently, 13/13 tests pass.